requirements.txt
base58==0.2.3
boto3==1.4.0
botocore==1.4.47
click==6.6
click-didyoumean==0.0.3
click-plugins==1.0.3
docutils==0.12
futures==3.0.5
hjson==2.0.0
jmespath==0.9.0
kappa==0.6.0
lambda-packages==0.8.0
munch==2.0.4
python-dateutil==2.5.3
python-slugify==1.2.0
requests>=2.10.0
six==1.10.0
tqdm==4.8.3
Werkzeug==0.11.10
wheel==0.29.0
wsgi-request-logger==0.4.5
